base node is clean now
